POEMTP - 1st Release
Hello everyone,

This script is what I Call POEMTP. It pretty much sends a e-mail with
the buying message someone sends you on POE (pretty much it gets
triggered when soneone sends a message that contains the words: buy, wtb
and the sentence "Hi, I would like to buy").

You'll need the newest version of python and also email.mime and smtplib
and of course a g-mail account ( in the future i'll add more options).

Keep in mind that this is the first version and the code far from what
it should be... But hey it works!

Remember to change lines 6,7 from _main_.py to the credentials that you
will be using to send the e-mail and at line 8 you should put the
address you want to receive the e-mail (You can use the same e-mail for
both sending and receiving. I personally created a new gmail account
just to sendind e-mails).

ps: Also check the address of your game folder on the 2nd line of
poegrep.py file.